django_test/requirements.txt
Django==3.1.7
django-filter==2.4.0
djangorestframework==3.12.3
gunicorn==20.0.4
psycopg2==2.8.5
